refactor : stream() API를 활용한 리팩터링
* List 컬렉션 -> Map -> 새로운 List객체의 생성

+++ src/main/java/lotto/domain/LottoRankRecord.java
@@ -5,7 +5,7 @@ import java.util.Objects;
 public class LottoRankRecord {
 
     private final LottoRank lottoRank;
-    private final int number;
+    private final long number;
 
     public LottoRankRecord(LottoRank lottoRank, int number) {
         this.lottoRank = lottoRank;
